[REFACTOR] suppression : ajout d'un dialog de confirmation lors de la suppression

+++ src/app/pages/parametres/employes/employes.component.ts
@@ -1,9 +1,9 @@
-import {Component} from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {Ferme} from '../../../models/ferme';
 import {FormBuilder, FormGroup, ReactiveFormsModule, Validators} from '@angular/forms';
 import {FermeService} from '../../../services/ferme.service';
 import {EmployeRequest} from '../../../models/ferme/employe-request';
-import {MessageService} from 'primeng/api';
+import {ConfirmationService, MessageService} from 'primeng/api';
 import {User} from '../../../models/user';
 import { CardModule} from 'primeng/card';
 import {NgForOf} from '@angular/common';
@@ -11,6 +11,7 @@ import {Button} from 'primeng/button';
 import {FloatLabel} from 'primeng/floatlabel';
 import {InputText} from 'primeng/inputtext';
 import { ScrollPanelModule } from 'primeng/scrollpanel';
+import {ConfirmDialogModule} from 'primeng/confirmdialog';
 
 @Component({
   selector: 'app-employes',
@@ -22,17 +23,18 @@ import { ScrollPanelModule } from 'primeng/scrollpanel';
     InputText,
     ReactiveFormsModule,
     ScrollPanelModule,
+    ConfirmDialogModule
   ],
   templateUrl: './employes.component.html',
   styleUrl: './employes.component.css'
 })
-export class EmployesComponent {
+export class EmployesComponent implements OnInit {
   employeForm: FormGroup;
   loading: boolean = false;
   employes: User[] = [];
 
   constructor(private formBuilder: FormBuilder, private fermeService: FermeService,
-              private messageService: MessageService) {
+              private messageService: MessageService, private confirmationService: ConfirmationService) {
     this.employeForm = this.formBuilder.group({
       email: ['', [Validators.required, Validators.email]],
       first_name: ['', [Validators.required]],
@@ -52,23 +54,41 @@ export class EmployesComponent {
     )
   }
 
-  onDeleteEmploye(employe: User) {
+  onDeleteEmploye(event: any, employe: User) {
     this.loading = true;
-    this.fermeService.deleteEmploye(employe.id).subscribe(
-      {
-        next: () => {
-          this.updateFerme();
-          this.messageService.add({
-            severity: 'success',
-            summary: 'Succès',
-            detail: `L'employé ${employe.first_name} a bien été supprimé.`
+    this.confirmationService.confirm({
+      target: event.target as EventTarget,
+      message: `Êtes vous sûr de vouloir supprimer l'employé '${employe.first_name} ${employe.last_name}' ?`,
+      header: 'Suppression',
+      icon: 'pi pi-info-circle',
+      rejectLabel: 'Annuler',
+      rejectButtonProps: {
+        label: 'Annuler',
+        severity: 'secondary',
+        outlined: true,
+      },
+      acceptButtonProps: {
+        label: 'Supprimer',
+        severity: 'danger',
+      },
+      accept: () => {
+        this.fermeService.deleteEmploye(employe.id).subscribe(
+          {
+            next: () => {
+              this.updateFerme();
+              this.messageService.add({
+                severity: 'success',
+                summary: 'Succès',
+                detail: `L'employé '${employe.first_name} ${employe.last_name}' a bien été supprimé.`
+              });
+              this.loading = false;
+            },
+            error: () => {
+              this.loading = false;
+            }
           });
-          this.loading = false;
-        },
-        error: () => {
-          this.loading = false;
-        }
-      });
+      }
+    });
   }
 
   onSubmit() {
